Simplest Formatting Working with STL
With the functionality added in C++20, many of the features added by the custom string view and span are present in the STL, so the plan is to switch to the standard implementations to save from having extra types. This commit is the first which covers the simplest formatting case.